Divide classes into packages
Currently the classes do not belong to any packages.

Lets group the classes into 2 different packages under the duke
package, the command and task package.

I also extracted out methods in Duke class that pertains to commands
and made a new class called CommandManager.

I also updated the runtest.bat file to ensure that compile path is
correct.
