+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public enum Command {
     LIST, DONE, TODO, DEADLINE, EVENT, BYE
 }
