+++ src/main/java/duke/command/IllegalCommandException.java
@@ -1,2 +1,4 @@
+package duke.command;
+
 public class IllegalCommandException extends Exception {
 }
