+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Deadline extends Task {
     private static final String DEADLINE_TAG = "[D]";
     private static final String BY_OPEN_TEXT = " (by: ";
