+++ src/main/java/duke/task/MissingTaskLiteralException.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class MissingTaskLiteralException extends Exception {
     public MissingTaskLiteralException(String literal) {
         super(literal);
